api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["wpcasarin <72427968+wpcasarin@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pypandoc-binary = "^1.14"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
python-multipart = "^0.0.17"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
